requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jpype1==1.6.0
konlpy==0.6.0
lxml==6.0.2
numpy==2.4.4
openai==2.31.0
packaging==26.0
pefile==2024.8.26
pillow==12.2.0
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
PyQt6==6.11.0
PyQt6-Qt6==6.11.0
PyQt6_sip==13.11.1
pystray==0.19.5
python-dotenv==1.2.2
pywin32-ctypes==0.2.3
RapidFuzz==3.14.5
regex==2026.4.4
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
wheel==0.46.3
